chore: move over to new tiles hdr
