+++ src/app/fdctl/run/run.h
@@ -2,7 +2,7 @@
 #define HEADER_fd_src_app_fdctl_run_h
 
 #include "../fdctl.h"
-#include "tiles/tiles.h"
+#include "../../../disco/tiles.h"
 #include "topos/topos.h"
 
 #include "../../../waltz/xdp/fd_xsk.h"
