+++ src/app/fdctl/run/tiles/fd_blackhole.c
@@ -1,4 +1,4 @@
-#include "tiles.h"
+#include "../../../../disco/tiles.h"
 
 /**
  * Tile used to filter out all packets when used as a consumer.
