+++ src/app/fdctl/run/tiles/fd_gossip.c
@@ -2,7 +2,7 @@
 
 #define _GNU_SOURCE 
 
-#include "tiles.h"
+#include "../../../../disco/tiles.h"
 
 #include <unistd.h>
 #include <arpa/inet.h>
