+++ src/app/fdctl/run/tiles/fd_pack.c
@@ -1,4 +1,4 @@
-#include "tiles.h"
+#include "../../../../disco/tiles.h"
 
 #include "generated/pack_seccomp.h"
