+++ src/app/fdctl/run/tiles/fd_replay_thread.c
@@ -1,3 +1,3 @@
-#include "tiles.h"
+#include "../../../../disco/tiles.h"
 
 fd_topo_run_tile_t fd_tile_replay_thread = { .name = "thread", .for_tpool = 1 };
