+++ src/app/fdctl/run/tiles/fd_sign.c
@@ -1,11 +1,12 @@
 #define _GNU_SOURCE
-#include "tiles.h"
+#include "../../../../disco/tiles.h"
 
 #include "generated/sign_seccomp.h"
 
 #include "../../../../disco/keyguard/fd_keyguard.h"
 #include "../../../../disco/keyguard/fd_keyload.h"
 
+#include <errno.h>
 #include <sys/mman.h>
 
 #define MAX_IN (32UL)
