+++ src/app/fdctl/run/topos/fd_frankendancer.c
@@ -1,4 +1,5 @@
-#include "../tiles/tiles.h"
+#include "../../../../disco/tiles.h"
+#include "../../fdctl.h"
 #include "../../config.h"
 #include "../../../../ballet/shred/fd_shred.h"
 #include "../../../../disco/topo/fd_topob.h"
