+++ src/app/fddev/tiles/fd_benchg.c
@@ -1,4 +1,4 @@
-#include "../../fdctl/run/tiles/tiles.h"
+#include "../../../disco/tiles.h"
 
 #include "../../../flamenco/types/fd_types_custom.h"
 #include "../../../flamenco/runtime/fd_system_ids_pp.h"
