+++ src/app/fddev/tiles/fd_bencho.c
@@ -1,4 +1,4 @@
-#include "../../fdctl/run/tiles/tiles.h"
+#include "../../../disco/tiles.h"
 
 #include "../rpc_client/fd_rpc_client.h"
 #include "../rpc_client/fd_rpc_client_private.h"
