+++ src/app/fddev/tiles/fd_benchs.c
@@ -1,11 +1,12 @@
 /* _GNU_SOURCE for recvmmsg and sendmmsg */
 #define _GNU_SOURCE
 
-#include "../../fdctl/run/tiles/tiles.h"
+#include "../../../disco/tiles.h"
 #include "../../../waltz/xdp/fd_xsk_aio.h"
 #include "../../../waltz/quic/fd_quic.h"
 #include "../../../waltz/tls/test_tls_helper.h"
 
+#include <errno.h>
 #include <linux/unistd.h>
 #include <sys/types.h>
 #include <sys/socket.h>
